Use hand rolled client, rather than generated
